frontend/ws_bridge/requirements-dev.txt
# Phase 4 bridge test harness: httpx.AsyncClient + pytest_asyncio + httpx-ws
# (TestClient + asyncio.new_event_loop binds fakeredis to the wrong loop;
# see plan section "Test harness convention for Tasks 7-10").
#
# Upper-bound pin on httpx-ws: 0.8.0 introduced an API change to
# ASGIWebSocketTransport (aconnect_ws now reads transport._task_group,
# which is only set when the transport is entered as an async context
# manager — our fixtures construct it directly). 0.7.2 is the version
# we developed against. Re-evaluate this pin when migrating the test
# harness off the private API (see TODOS.md "Extract bridge WS test
# helpers to conftest.py").
httpx-ws>=0.7.0,<0.8

# Hang protection: pytest.ini sets timeout = 30 (eng-review 4A). Without
# this dep, the timeout setting is silently ignored.
pytest-timeout>=2.3
